product features updated

+++ database/migrations/2023_08_18_075824_create_guarantees_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->string('guarantee_name');
             $table->unsignedBigInteger('product_id')->nullable();
             $table->foreign('product_id')->references('id')->on('products')->onDelete('cascade');
-            $table->decimal('price_increase',20,1)->default(0);
+            $table->unsignedBigInteger('price_increase')->default(0);
             $table->tinyInteger('status')->default(0);
             $table->softDeletes();
             $table->timestamps();
